Crud is possible; Added support for delete on both entities; Fixed parameters for beacons and sensor; Added function to return all data in json; Need to return it from another file;

+++ main.php
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+//session_start();
 /* setting the default timezone to Copenhagen/denmark */
 date_default_timezone_set('Europe/Copenhagen');
 
@@ -59,12 +59,14 @@ foreach($_GET as $key=>$value){
 $entityType = $_GET['entype']; // int - specifies entity type #1 = Beacon; #2 = Sensor
 $entity = null; // var for defining sensor or beacon
 // define sensor params
-$id = $_GET['id']; // id for both entities
+
+//$id = "key" . 3;
+$id = "key" . mt_rand(100000,999999); ; // id for both entities
 $sensorType = $_GET['sensortype']; // Type
 $value = $_GET['value']; // value for both
 $timestamp = $_GET['timestamp'];
 $timestampInject = date("Y-m-d H:i:s");
-$androidID = $_GET['androidID']; // value for both
+//$androidID = $_GET['androidID']; // value for both
 // define beacon params
 $lat = $_GET['lat'];
 $long = $_GET['long'];
@@ -76,38 +78,55 @@ $UUID = $_GET['uuid'];
 // only for testing
 $deleteID = $_GET['del'];
 $updateID = $_GET['upd'];
+$GetData = $_GET['getdata'];
+
+$obj_store_fetch = new GDS\Store('Beacon');
+$obj_store_fetch2 = new GDS\Store('Sensor');
+
 
 if($entityType == 1) {
     $entity = "Beacon";
+
+    //$query = $obj_store_fetch->fetchAll();
+    $query = $obj_store_fetch->fetchAll("select * from Beacon where UUID = '".$UUID."' AND major = '".$major."' AND minor= '".$minor."'");
+    echo PHP_EOL, "Query found ", count($query), ":Beacon records", PHP_EOL . "<br>";
+
+    if(!$query){
+    echo "worked";
     echo "entityType: " . $entity;
 
     echo '<p/>ID: ', $id, "<br/>";
 
     $obj_1 = new GDS\Entity();
-    //$obj_1->id = $id;
+    $obj_1->Bid = $id;
     $obj_1->lat = $lat;
     $obj_1->long = $long;
     $obj_1->beaconID = $beaconID;
     $obj_1->major = $major;
     $obj_1->minor = $minor;
-    $obj_1->androidID = $androidID;
+    //$obj_1->androidID = $androidID;
     $obj_1->UUID = $UUID;
     // Write it to Datastore
     $obj_store = new GDS\Store($entity);
     $obj_store->upsert($obj_1);
 
     $obj_store_fetch = new GDS\Store($entity);
+    }
+    else{
+        echo "already exists";
+    }
 }
 else if($entityType == 2){
     $entity = "Sensor";
     echo "entityType: " . $entity;
+
     //echo '<a class="clearurl" href="\?entype=2&id=1&sensortype=sensor1&value=723&timestamp='.$timestampInject.'&androidID=ad1">add a sensor entity</a><br>'
     $obj_2 = new GDS\Entity();
     //$obj_2->id = $id;
     $obj_2->sensortype = $sensorType;
     $obj_2->value = $value;
     $obj_2->timestamp = $timestamp;
-    $obj_2->androidID = $androidID;
+    $obj_2->Sid = $id;
     // Write it to Datastore
     $obj_store = new GDS\Store($entity);
     $obj_store->upsert($obj_2);
@@ -148,8 +167,7 @@ else {
 
 
 
-$obj_store_fetch = new GDS\Store('Beacon');
-$obj_store_fetch2 = new GDS\Store('Sensor');
+
 
 //show($obj_store_fetch->fetchAll());
 
@@ -162,7 +180,7 @@ function show($arr)
 {
     echo PHP_EOL, "Query found ", count($arr), ":Beacon records", PHP_EOL . "<br>";
     foreach ($arr as $obj_model) {
-        echo "   ID: {$obj_model->ID}, beaconID: {$obj_model->beaconID}", PHP_EOL . "<br>";
+        echo "   UUID: {$obj_model->UUID}, major: {$obj_model->major},minor: {$obj_model->minor},beaconID: {$obj_model->beaconID},BID: {$obj_model->Bid}", PHP_EOL . "<br>";
     }
 }
 
@@ -172,29 +190,37 @@ function show2($arr)
     echo PHP_EOL, "Query found ", count($arr), ":sensor records", PHP_EOL . "<br>";
     foreach ($arr as $obj_model) {
 
-        echo "   ID: {$obj_model->ID}, value: {$obj_model->value}", PHP_EOL . "<br>";
+        echo "  sensorType: {$obj_model->sensortype},Timestamp: {$obj_model->timestamp},SID: {$obj_model->Sid}", PHP_EOL . "<br>";
 
 
     }
 }
 
+//show($obj_store_fetch->fetchAll("select * from Beacon where Bid=key387673"));
     show($obj_store_fetch->fetchAll());
     show2($obj_store_fetch2->fetchAll());
+//show2($obj_store_fetch2->fetchAll("select * from Sensor where value=1"));
 
 
 
 /* delete script */
 if($deleteID == 1){
-    $arr_1 = $obj_store_fetch->fetchAll();
+    //$arr_1 = $obj_store_fetch->fetchAll();
+    $arr_1 = $obj_store_fetch->fetchOne("select * from Beacon where Bid='".$updateID."'"); // deleting based on GQL
     echo "Found ", count($arr_1), " records", PHP_EOL;
-    $obj_store_fetch->delete($arr_1);
-    echo "deleted all beacon entities!";
+    if($arr_1 != 0){
+        $obj_store_fetch->delete($arr_1);
+        echo "deleted beacon entity with id:" .$updateID ;
+    }
+    else{
+        echo "no id found";
+    }
 }
 else if($deleteID == 2){
 
     //$arr_1 = $obj_store_fetch2->fetchById("4967730973245440");
     //$arr_1 = $obj_store_fetch2->fetchAll();
-    $arr_1 = $obj_store_fetch2->fetchOne("select * from Sensor where value=725"); // deleting based on GQL
+    $arr_1 = $obj_store_fetch2->fetchOne("select * from Sensor where Sid='".$updateID."'"); // deleting based on GQL
     echo "Found ", count($arr_1), " records", PHP_EOL;
     if($arr_1 != 0){
         $obj_store_fetch2->delete($arr_1);
@@ -207,12 +233,38 @@ else if($deleteID == 2){
 
 }
 /* update */
-if($updateID){
+/*if($updateID){
     $arr_1 = $obj_store_fetch2->fetchOne("select * from Sensor where value='".$updateID."'"); // selecting based on GQL
     echo "Found ", count($arr_1), " records", PHP_EOL;
 }
 else{
     echo "no entity with that id found";
+}*/
+/* return all data as json to contextApp */
+if($GetData ==1){
+    $items = array();
+   $arr =  $obj_store_fetch->fetchAll();
+    foreach ($arr as $obj_model) {
+        $items[] = $obj_model->getData();
+
+        //print_r(array_values($items));
+        echo json_encode($items, JSON_PRETTY_PRINT);
+
+
+    }
+}
+else if($GetData == 2){
+
+    $items = array();
+    $arr =  $obj_store_fetch2->fetchAll();
+    foreach ($arr as $obj_model) {
+        $items[] = $obj_model->getData();
+
+        //print_r(array_values($items));
+        echo json_encode($items, JSON_PRETTY_PRINT);
+
+
+    }
 }
 
 ?>
@@ -223,41 +275,53 @@ else{
 </head>
 <body>
 <hr>
-<button><a href="\?del=1">delete beacon</a></button><br>
+<button><a id="delbeacon" href="\?del=1">delete beacon</a></button><br>
 <button><a id="delsensor" href="\?del=2">delete sensor</a></button><br>
 <button><a id="update" href="">update sensor</a></button><br>
 
 
 
-<button><a class="clearurl" href="\?entype=1&id=1&lat=x234&long=y723&beacon=bID1&minor=m47&major=m67&androidID=ad1&uuid=uniqid1">add a beacon entity</a></button><br>
+<button><a id="addbeacon" class="clearurl" href="\?entype=1&id=1&lat=x234&long=y723&minor=m47&major=m67&uuid=uniqid1">add a beacon entity</a></button><br>
 <button>
 <?php
-echo '<a class="clearurl" href="\?entype=2&id=1&sensortype=sensor1&timestamp='.$timestampInject.'&androidID=ad1&value=755">add a sensor entity</a><br>';
+echo '<a id="addsensor" class="clearurl" href="\?entype=2&id=1&sensortype=sensor1&timestamp='.$timestampInject.'&androidID=ad1">add a sensor entity</a><br>';
 ?>
     </button>
 <br>
-<label for="val">write value to update/delete</label>
+<label for="val">write value to update/delete. EntityID:</label>
 <input type="text" id="val">
+<span id="chosenID">ID:</span>
 
 <script>
     jQuery(document).ready(function($){
         /* clear url on refresh */
-        var url = window.location.href;
-        url = url.split( '?' );
-        if(url[1] == null){
-            //alert("url is clear");
-        }
-        else{
-            //alert("url contains params");
-            window.location.href = url[0];
-        }
+        //refreshPage();
 
         var val =  $("#val").val();
         $("#val").change(function(){
             $("#update").attr("href", "\?upd="+$("#val").val());
+            $("#delbeacon").attr("href", "\?upd="+$("#val").val()+"&del=1");
             $("#delsensor").attr("href", "\?upd="+$("#val").val()+"&del=2");
-            alert("changed" + $("#val").val());
+            var hrefs = $('#addsensor').attr('href');
+            var hrefb = $('#addbeacon').attr('href');
+            $("#addsensor").attr("href", hrefs + "&value=" + $("#val").val());
+            $("#addbeacon").attr("href", hrefb + "&beacon=" + $("#val").val());
+            //&beacon=bID1
+            $("chosenID").html("ID:" + $("#val").val());
+            //alert("changed" + $("#val").val());
         });
+
+        function refreshPage(){
+            var url = window.location.href;
+            url = url.split( '?' );
+            if(url[1] == null){
+                //alert("url is clear");
+            }
+            else{
+                //alert("url contains params");
+                window.location.href = url[0];
+            }
+        }
     });
 </script>
 </body>
